have implement with viewPager2 and the item history in city activity

+++ app/src/main/java/com/example/weatherappsample1yt/data/model/weatherAPI/ForecastResponseApiWA.kt
@@ -48,7 +48,7 @@ data class ForecastResponseApiWA(
         @SerializedName("pressure_mb")
         val pressureMb: Int?,
         @SerializedName("temp_c")
-        val tempC: Int?,
+        val tempC: Double?,
         @SerializedName("temp_f")
         val tempF: Double?,
         @SerializedName("uv")
